+++ History.md
@@ -4,6 +4,11 @@ Release date: unreleased
 * Dropped support for Ruby 2.7, 3.0+ is now required
 * Dropped support for Selenium < 4.8
 
+### Added
+
+* `Capybara::Minitest::Assertions.{asserts,asserts_no}` and
+  `Capybara::Minitest::Assertions.{asserts_selector,asserts_no_selector}` [Sean Doyle]
+
 # Version 3.39.2
 Release date: 2023-06-10